#695: Actually, use integers for smallint columns.
That takes away a little more space in the database, but I'm willing to sacrifice that in order to prevent possible bugs in the future, introduced by someone who thinks Laravel's "boolean" needs to be a true boolean. It isn't right now, but it still looks confusing in the migration file.